feat: add api gateway to front lambda to get counter
