+++ terraform/stage/main.tf
@@ -23,7 +23,6 @@ provider "aws" {
 locals {
   account_id       = "973533177904"
   region           = "eu-west-2"
-  lambda_zip_path  = "../../lambda.zip"
   env              = "stage"
   s3_origin_id     = "CVBucketOrigin"
   root_domain_name = "bernardting.com"
